[Refactoring] Transfer all DAO tests in correct dir
