+++ src/test/java/ua/strongBody/dao/impl/CartDAOImplIntegrationTest.java
@@ -1,4 +1,4 @@
-package ua.strongBody.dao;
+package ua.strongBody.dao.impl;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
